+++ nym-vpn-core/crates/nym-vpn-account-controller/src/commands/request_zknym/mod.rs
@@ -1,0 +1,13 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+mod cached_data;
+mod error;
+mod handler;
+mod request;
+
+pub use error::RequestZkNymError;
+pub use handler::RequestZkNymSummary;
+pub use request::RequestZkNymSuccess;
+
+pub(crate) use handler::{WaitingRequestZkNymCommandHandler, ZkNymId};
