+++ nym-vpn-core/crates/nym-vpn-account-controller/src/commands/request_zknym/request.rs
@@ -1,12 +1,8 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
 use std::{
     collections::HashMap,
-    sync::{
-        atomic::{AtomicU32, Ordering},
-        Arc,
-    },
     time::{Duration, Instant},
 };
 
@@ -25,321 +21,25 @@ use nym_vpn_api_client::{
 };
 use serde::{Deserialize, Serialize};
 use time::Date;
-use tokio::task::JoinSet;
 
 use crate::{
     commands::VpnApiEndpointFailure,
-    shared_state::RequestZkNymResult,
     storage::{PendingCredentialRequest, VpnCredentialStorage},
-    SharedAccountState,
 };
 
-use super::{AccountCommandError, AccountCommandResult};
-
-// The maximum number of zk-nym requests that can fail in a row before we disable background
-// refresh
-const ZK_NYM_MAX_FAILS: u32 = 10;
+use super::{cached_data::CachedData, RequestZkNymError, ZkNymId};
 
 const ZK_NYM_POLLING_TIMEOUT: Duration = Duration::from_secs(60);
 const ZK_NYM_POLLING_INTERVAL: Duration = Duration::from_secs(5);
 
-#[derive(Clone, Default)]
-struct CachedData {
-    partial_verification_keys:
-        Arc<tokio::sync::Mutex<HashMap<u64, PartialVerificationKeysResponse>>>,
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RequestZkNymSuccess {
+    pub id: ZkNymId,
 }
 
-impl CachedData {
-    async fn get_partial_verification_keys(
-        &self,
-        epoch_id: u64,
-        vpn_api_client: &VpnApiClient,
-    ) -> Result<PartialVerificationKeysResponse, RequestZkNymError> {
-        // Get the partial verification keys for the given epoch if they exist in the cache.
-        // Otherwise fetch it from the API, store it and then return it
-        let mut partial_verification_keys = self.partial_verification_keys.lock().await;
-        if let Some(issuers) = partial_verification_keys.get(&epoch_id) {
-            tracing::debug!("Using cached partial verification keys for epoch: {epoch_id}");
-            Ok(issuers.clone())
-        } else {
-            tracing::info!("Fetching partial verification keys for epoch: {epoch_id}");
-            let issuers = vpn_api_client
-                .get_directory_zk_nyms_ticketbook_partial_verification_keys()
-                .await
-                .map_err(|err| {
-                    VpnApiEndpointFailure::try_from(err)
-                        .map(|source| {
-                            RequestZkNymError::GetPartialVerificationKeysEndpointFailure {
-                                source,
-                                epoch_id,
-                            }
-                        })
-                        .unwrap_or_else(RequestZkNymError::unexpected_response)
-                })?;
-
-            if issuers.epoch_id != epoch_id {
-                return Err(RequestZkNymError::EpochIdMismatch);
-            }
-
-            partial_verification_keys.insert(epoch_id, issuers.clone());
-            Ok(issuers)
-        }
-    }
-}
-
-pub(crate) struct WaitingRequestZkNymCommandHandler {
-    credential_storage: VpnCredentialStorage,
-    account_state: SharedAccountState,
-    vpn_api_client: VpnApiClient,
-    zk_nym_fails_in_a_row: Arc<AtomicU32>,
-
-    // Cache some of the data used to import zk-nyms between requests, to speed things up. Consider
-    // persisting this to storage
-    cached_data: CachedData,
-}
-
-impl WaitingRequestZkNymCommandHandler {
-    pub(crate) fn new(
-        credential_storage: VpnCredentialStorage,
-        account_state: SharedAccountState,
-        vpn_api_client: nym_vpn_api_client::VpnApiClient,
-    ) -> Self {
-        WaitingRequestZkNymCommandHandler {
-            credential_storage,
-            account_state,
-            vpn_api_client,
-            zk_nym_fails_in_a_row: Default::default(),
-            cached_data: Default::default(),
-        }
-    }
-
-    pub(crate) fn build(
-        &self,
-        account: VpnApiAccount,
-        device: Device,
-    ) -> RequestZkNymCommandHandler {
-        let id = uuid::Uuid::new_v4();
-        tracing::debug!("Created new zk-nym request command handler: {}", id);
-        RequestZkNymCommandHandler {
-            id,
-            account,
-            device,
-            credential_storage: self.credential_storage.clone(),
-            account_state: self.account_state.clone(),
-            vpn_api_client: self.vpn_api_client.clone(),
-            zk_nym_fails_in_a_row: self.zk_nym_fails_in_a_row.clone(),
-            cached_data: self.cached_data.clone(),
-        }
-    }
-
-    pub(crate) fn reset(&self) {
-        self.zk_nym_fails_in_a_row.store(0, Ordering::Relaxed);
-    }
-
-    pub(crate) async fn max_fails_reached(&self) -> bool {
-        self.zk_nym_fails_in_a_row.load(Ordering::Relaxed) >= ZK_NYM_MAX_FAILS
-    }
-}
-
-pub(crate) struct RequestZkNymCommandHandler {
-    id: uuid::Uuid,
-    account: VpnApiAccount,
-    device: Device,
-    credential_storage: VpnCredentialStorage,
-    account_state: SharedAccountState,
-    vpn_api_client: VpnApiClient,
-
-    zk_nym_fails_in_a_row: Arc<AtomicU32>,
-    cached_data: CachedData,
-}
-
-impl RequestZkNymCommandHandler {
-    fn id_str(&self) -> String {
-        format!("{:.8}", self.id.to_string())
-    }
-
-    pub(crate) async fn run(self) -> AccountCommandResult {
-        AccountCommandResult::RequestZkNym(self.request_zk_nyms_outer().await)
-    }
-
-    #[tracing::instrument(
-        skip(self),
-        fields(id = %self.id_str()),
-        ret,
-        err,
-    )]
-    async fn request_zk_nyms_outer(self) -> Result<RequestZkNymSummary, AccountCommandError> {
-        tracing::debug!("Running zk-nym request command handler: {}", self.id);
-
-        // Defensive check for something that should not be possible
-        if self.account_state.is_zk_nym_request_in_progress().await {
-            return Err(AccountCommandError::internal(
-                "duplicate zk-nym request command",
-            ));
-        }
-
-        self.account_state
-            .set_zk_nym_request(RequestZkNymResult::InProgress)
-            .await;
-
-        match self.request_zk_nyms().await {
-            Ok(success) => {
-                self.account_state
-                    .set_zk_nym_request(RequestZkNymResult::from(success.clone()))
-                    .await;
-                Ok(success)
-            }
-            Err(err) => {
-                self.account_state
-                    .set_zk_nym_request(RequestZkNymResult::from(err.clone()))
-                    .await;
-                Err(AccountCommandError::from(err))
-            }
-        }
-    }
-
-    #[tracing::instrument(
-        skip(self),
-        fields(id = %self.id_str()),
-        ret,
-        err,
-    )]
-    async fn request_zk_nyms(&self) -> Result<RequestZkNymSummary, RequestZkNymError> {
-        tracing::debug!("Running zk-nym request command handler: {}", self.id);
-
-        // If we have pending zk-nym ticketbooks, try those first
-        let resumed_requests = self.resume_request_zk_nyms().await;
-
-        let ticket_types = self.check_ticket_types_running_low().await?;
-        tracing::debug!("Ticket types running low: {:?}", ticket_types);
-
-        let new_requests = self.request_zk_nyms_for_ticket_types(ticket_types).await;
-
-        let zk_nym_fails_in_a_row = self.zk_nym_fails_in_a_row.load(Ordering::Relaxed);
-        if zk_nym_fails_in_a_row > 0 {
-            tracing::warn!("We have reached {zk_nym_fails_in_a_row} zk-nym fails in a row",);
-        }
-
-        let result = resumed_requests
-            .into_iter()
-            .chain(new_requests.into_iter())
-            .collect();
-
-        Ok(result)
-    }
-
-    async fn check_ticket_types_running_low(&self) -> Result<Vec<TicketType>, RequestZkNymError> {
-        self.credential_storage
-            .check_ticket_types_running_low()
-            .await
-            .map_err(RequestZkNymError::internal)
-    }
-
-    async fn request_zk_nyms_for_ticket_types(
-        &self,
-        ticket_types: Vec<TicketType>,
-    ) -> Vec<Result<RequestZkNymSuccess, RequestZkNymError>> {
-        tracing::info!("Requesting zk-nym ticketbooks for: {:?}", ticket_types);
-
-        let account = self.account.clone();
-        let device = self.device.clone();
-        let vpn_api_client = self.vpn_api_client.clone();
-        let credential_storage = self.credential_storage.clone();
-        let cached_data = self.cached_data.clone();
-
-        let mut join_set = JoinSet::new();
-        for ticket_type in ticket_types {
-            join_set.spawn(request_zk_nym(
-                ticket_type,
-                account.clone(),
-                device.clone(),
-                vpn_api_client.clone(),
-                credential_storage.clone(),
-                cached_data.clone(),
-            ));
-        }
-        wait_for_join_set(join_set).await
-    }
-
-    async fn resume_request_zk_nyms(&self) -> Vec<Result<RequestZkNymSuccess, RequestZkNymError>> {
-        let to_resume = self
-            .check_zk_nyms_possible_to_resume()
-            .await
-            .inspect_err(|err| {
-                tracing::error!("Failed to check zk-nyms possible to resume: {:?}", err);
-            })
-            .unwrap_or_default();
-        self.resume_request_zk_nyms_for_ids(to_resume).await
-    }
-
-    async fn check_zk_nyms_possible_to_resume(&self) -> Result<Vec<ZkNymId>, RequestZkNymError> {
-        let zk_nyms_available_for_download = self.get_zk_nyms_available_for_download().await?;
-
-        self.credential_storage
-            .clean_up_stale_requests()
-            .await
-            .inspect_err(|err| {
-                tracing::error!("Failed to clean up stale requests: {:?}", err);
-            })
-            .ok();
-
-        let pending_requests_data = self
-            .credential_storage
-            .get_pending_request_ids()
-            .await
-            .map_err(RequestZkNymError::internal)?;
-
-        let zk_nyms_possible_to_resume = zk_nyms_available_for_download
-            .into_iter()
-            .filter(|zk_nym| pending_requests_data.contains(zk_nym))
-            .collect();
-
-        Ok(zk_nyms_possible_to_resume)
-    }
-
-    async fn get_zk_nyms_available_for_download(&self) -> Result<Vec<ZkNymId>, RequestZkNymError> {
-        self.vpn_api_client
-            .get_zk_nyms_available_for_download(&self.account, &self.device)
-            .await
-            .map(|response| response.items.into_iter().map(|item| item.id).collect())
-            .map_err(|err| {
-                VpnApiEndpointFailure::try_from(err)
-                    .map(
-                        |source| RequestZkNymError::GetZkNymsAvailableForDownloadEndpointFailure {
-                            source,
-                        },
-                    )
-                    .unwrap_or_else(RequestZkNymError::unexpected_response)
-            })
-    }
-
-    async fn resume_request_zk_nyms_for_ids(
-        &self,
-        pending_requests: Vec<ZkNymId>,
-    ) -> Vec<Result<RequestZkNymSuccess, RequestZkNymError>> {
-        if pending_requests.is_empty() {
-            return Vec::new();
-        }
-        tracing::info!("Resuming {} zk-nym requests", pending_requests.len());
-
-        let account = self.account.clone();
-        let device = self.device.clone();
-        let vpn_api_client = self.vpn_api_client.clone();
-        let credential_storage = self.credential_storage.clone();
-        let cached_data = self.cached_data.clone();
-
-        let mut join_set = JoinSet::new();
-        for pending_request in pending_requests {
-            join_set.spawn(resume_request_zk_nym(
-                pending_request,
-                account.clone(),
-                device.clone(),
-                vpn_api_client.clone(),
-                credential_storage.clone(),
-                cached_data.clone(),
-            ));
-        }
-        wait_for_join_set(join_set).await
+impl RequestZkNymSuccess {
+    pub fn new(id: ZkNymId) -> Self {
+        RequestZkNymSuccess { id }
     }
 }
 
@@ -353,7 +53,7 @@ pub(crate) struct ZkNymRequestData {
 }
 
 #[tracing::instrument(skip(account, device, vpn_api_client, credential_storage, cached_data))]
-async fn request_zk_nym(
+pub(super) async fn request_zk_nym(
     ticketbook_type: TicketType,
     account: VpnApiAccount,
     device: Device,
@@ -386,7 +86,7 @@ async fn request_zk_nym(
 }
 
 #[tracing::instrument(skip(account, device, vpn_api_client, credential_storage, cached_data))]
-async fn resume_request_zk_nym(
+pub(super) async fn resume_request_zk_nym(
     id: ZkNymId,
     account: VpnApiAccount,
     device: Device,
@@ -889,199 +589,3 @@ async fn confirm_zk_nym_downloaded(
         })
         .inspect(|response| tracing::debug!("Confirmed zk-nym download: {}", response))
 }
-
-async fn wait_for_join_set(
-    mut join_set: JoinSet<Result<RequestZkNymSuccess, RequestZkNymError>>,
-) -> Vec<Result<RequestZkNymSuccess, RequestZkNymError>> {
-    let mut partial_results = Vec::new();
-    loop {
-        tokio::select! {
-            _ = tokio::time::sleep(Duration::from_secs(5 * 60)) => {
-                tracing::warn!("Request zk-nym timed out");
-                break;
-            }
-            result = join_set.join_next() => match result {
-                Some(Ok(result)) => {
-                    partial_results.push(result);
-                }
-                Some(Err(err)) => {
-                    tracing::error!("Failed to wait for task: {:?}", err);
-                }
-                None => {
-                    tracing::debug!("All zk-nym requests finished");
-                    break;
-                }
-            }
-        }
-    }
-    partial_results
-}
-
-pub(crate) type ZkNymId = String;
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RequestZkNymSuccess {
-    pub id: ZkNymId,
-}
-
-impl RequestZkNymSuccess {
-    pub fn new(id: ZkNymId) -> Self {
-        RequestZkNymSuccess { id }
-    }
-}
-
-#[derive(Clone, Debug, thiserror::Error, PartialEq, Eq, Serialize, Deserialize)]
-pub enum RequestZkNymError {
-    #[error("failed to get zk-nyms available for download: {source}")]
-    GetZkNymsAvailableForDownloadEndpointFailure { source: VpnApiEndpointFailure },
-
-    #[error("failed to create ecash keypair: {0}")]
-    CreateEcashKeyPair(String),
-
-    #[error("failed to construct withdrawal request: {0}")]
-    ConstructWithdrawalRequest(String),
-
-    #[error("failed to request zk-nym endpoint for {ticket_type}: {source}")]
-    RequestZkNymEndpointFailure {
-        ticket_type: String,
-        source: VpnApiEndpointFailure,
-    },
-
-    #[error("response contains invalid ticketbook type: {0}")]
-    InvalidTicketTypeInResponse(String),
-
-    #[error("ticket type mismatch")]
-    TicketTypeMismatch,
-
-    #[error("error polling for zknym result: {source}")]
-    PollZkNymEndpointFailure { source: VpnApiEndpointFailure },
-
-    #[error("polling task failed")]
-    PollingTaskError,
-
-    #[error("timeout polling for zknym {id}")]
-    PollingTimeout { id: ZkNymId },
-
-    #[error("polling for zknym {id} finished with error for ticket type: {ticket_type}")]
-    FinishedWithError {
-        id: ZkNymId,
-        ticket_type: String,
-        status: NymVpnZkNymStatus,
-    },
-
-    #[error("response is missing blinded shares")]
-    MissingBlindedShares,
-
-    #[error("response contains invalid master verification key: {0}")]
-    ResponseHasInvalidMasterVerificationKey(String),
-
-    #[error("epoch id mismatch")]
-    EpochIdMismatch,
-
-    #[error("expiration date mismatch")]
-    ExpirationDateMismatch,
-
-    #[error("failed to request partial verification keys for epoch {epoch_id}: {source}")]
-    GetPartialVerificationKeysEndpointFailure {
-        epoch_id: u64,
-        source: VpnApiEndpointFailure,
-    },
-
-    #[error("no master verification key in storage")]
-    NoMasterVerificationKeyInStorage,
-
-    #[error("no coin index signatures in storage")]
-    NoCoinIndexSignaturesInStorage,
-
-    #[error("no expiration date signatures in storage")]
-    NoExpirationDateSignaturesInStorage,
-
-    #[error("invalid verification key: {0}")]
-    InvalidVerificationKey(String),
-
-    #[error("failed to deserialize blinded signature: {0}")]
-    DeserializeBlindedSignature(String),
-
-    #[error("decoded keys missing index")]
-    DecodedKeysMissingIndex,
-
-    #[error("failed to import zknym")]
-    ImportZkNym { ticket_type: String, error: String },
-
-    #[error("failed to aggregate wallets: {0}")]
-    AggregateWallets(String),
-
-    #[error("failed to confirm zknym {id} download: {source}")]
-    ConfirmZkNymDownloadEndpointFailure {
-        id: ZkNymId,
-        source: VpnApiEndpointFailure,
-    },
-
-    #[error("missing pending request: {0}")]
-    MissingPendingRequest(ZkNymId),
-
-    #[error("failed to remove pending zk-nym request {id}: {error}")]
-    RemovePendingRequest { id: String, error: String },
-
-    #[error("credential storage error: {0}")]
-    CredentialStorage(String),
-
-    #[error("internal error: {0}")]
-    Internal(String),
-
-    #[error("unexpected error response: {0}")]
-    UnexpectedErrorResponse(String),
-}
-
-impl RequestZkNymError {
-    pub fn internal(message: impl ToString) -> Self {
-        RequestZkNymError::Internal(message.to_string())
-    }
-
-    pub fn unexpected_response(message: impl ToString) -> Self {
-        RequestZkNymError::UnexpectedErrorResponse(message.to_string())
-    }
-
-    pub fn message(&self) -> String {
-        match self {
-            RequestZkNymError::RequestZkNymEndpointFailure {
-                source,
-                ticket_type: _,
-            }
-            | RequestZkNymError::PollZkNymEndpointFailure { source } => source.message.clone(),
-            other => other.to_string(),
-        }
-    }
-
-    pub fn message_id(&self) -> Option<String> {
-        match self {
-            RequestZkNymError::RequestZkNymEndpointFailure {
-                source,
-                ticket_type: _,
-            }
-            | RequestZkNymError::PollZkNymEndpointFailure { source } => source.message_id.clone(),
-            _ => None,
-        }
-    }
-
-    pub fn ticket_type(&self) -> Option<String> {
-        match self {
-            RequestZkNymError::RequestZkNymEndpointFailure {
-                source: _,
-                ticket_type,
-            } => Some(ticket_type.clone()),
-            RequestZkNymError::FinishedWithError {
-                id: _,
-                ticket_type,
-                status: _,
-            }
-            | RequestZkNymError::ImportZkNym {
-                ticket_type,
-                error: _,
-            } => Some(ticket_type.clone()),
-            _ => None,
-        }
-    }
-}
-
-pub type RequestZkNymSummary = Vec<Result<RequestZkNymSuccess, RequestZkNymError>>;
